requirements.txt
langchain
langchain-openai
langchain-qdrant
langchain-community
langchain-neo4j
langchain-experimental
python-dotenv
neo4j
qdrant-client
openai
yt-dlp
